requirements.txt
asgiref==3.10.0
Django==5.2.8
django-environ==0.12.0
sqlparse==0.5.3
tzdata==2025.2
